use [[]]] style tests

+++ init/70_set_folders_and_config.sh
@@ -1,21 +1,13 @@
 #!/bin/bash
-mkdir -p /config/rtorrent /config/log/rtorrent
-if [ ! -f "/config/rtorrent/config.php" ]; then
-cp /defaults/config.php /config/rtorrent/config.php
-fi
-cp /config/rtorrent/config.php  /config/www/webui/conf/config.php
 
-if [ ! -f "/config/rtorrent/rtorrent.rc" ]; then
-cp /defaults/rtorrent.rc /config/rtorrent/rtorrent.rc
-fi
+mkdir -p /config/rtorrent/rtorrent_sess /config/log/rtorrent /detach_sess
 
-if [ ! -d "/config/rtorrent/rtorrent_sess" ]; then
-mkdir -p /config/rtorrent/rtorrent_sess
-fi
+[[ ! -f /config/rtorrent/config.php ] && cp /defaults/config.php /config/rtorrent/config.php
+[[ ! -L /config/www/webui/conf/config.php && -f /config/www/webui/conf/config.php ]] && rm /config/www/webui/conf/config.php
+[[ ! -L /config/www/webui/conf/config.php ]] && ln -s /config/rtorrent/config.php /config/www/webui/conf/config.php
 
-if [ ! -d "/detach_sess" ]; then
-mkdir -p /detach_sess
-fi
+
+[[ ! -f /config/rtorrent/rtorrent.rc ]] && cp /defaults/rtorrent.rc /config/rtorrent/rtorrent.rc
 
 chown abc:abc -R /config /detach_sess
 chown abc:abc /downloads
